Split metadata and page into 2 files
